requirements.txt
streamlit
openai
python-dotenv
jinja2
